requirements.txt
fastapi
uvicorn
python-multipart
pydantic
openai
langchain
faiss-cpu
pymupdf
pdfplumber
python-docx
streamlit
python-dotenv
numpy
reportlab
